+++ delphi/GrFN/__init__.py
@@ -1,1 +1,1 @@
-from delphi.translators.for2py.scripts.genPGM import create_pgm_dict, get_asts_from_files
+from delphi.translators.for2py.genPGM import create_pgm_dict, get_asts_from_files
