+++ delphi/translators/for2py/f2py_pp.py
@@ -20,7 +20,7 @@ Author:
 
 import sys
 from typing import List
-from delphi.translators.for2py.scripts.fortran_syntax import *
+from delphi.translators.for2py.fortran_syntax import *
 
 
 def process(lines: List[str]) -> str:
