+++ delphi/translators/for2py/for2py_arrays.py
