+++ delphi/translators/for2py/fortran_syntax.py
