+++ delphi/translators/for2py/genCode.py
