+++ delphi/translators/for2py/mod_index_generator.py
