+++ delphi/translators/for2py/translate.py
@@ -26,7 +26,7 @@ import pickle
 import re
 from collections import *
 import xml.etree.ElementTree as ET
-from delphi.translators.for2py.scripts.get_comments import (
+from delphi.translators.for2py.get_comments import (
     get_comments,
 )
 from typing import List, Dict
