+++ tests/data/arrays-basic-06.py
@@ -1,8 +1,8 @@
 import sys
 from typing import List
 import math
-from delphi.translators.scripts.for2py.fortran_format import *
-from delphi.translators.scripts.for2py.for2py_arrays import *
+from delphi.translators.for2py.fortran_format import *
+from delphi.translators.for2py.for2py_arrays import *
 from dataclasses import dataclass
