+++ tests/data/derived-types/Translation/derived-types-01.py
@@ -1,4 +1,4 @@
-from delphi.translators.for2py.scripts.fortran_format import *
+from delphi.translators.for2py.fortran_format import *
 import sys
 
 class mytype:
