+++ tests/data/derived-types/Translation/derived-types-02.py
@@ -1,6 +1,6 @@
 import sys
-from delphi.translators.for2py.scripts.fortran_format import *
-from delphi.translators.for2py.scripts.for2py_arrays import *
+from delphi.translators.for2py.fortran_format import *
+from delphi.translators.for2py.for2py_arrays import *
 
 class mytype:
     def __init__(self):
