+++ tests/data/derived-types/Translation/interface-01.py
@@ -1,4 +1,4 @@
-from delphi.translators.for2py.scripts.fortran_format import *
+from delphi.translators.for2py.fortran_format import *
 from M_mymod import *
 
 def main():
